+++ src/typings/cookie.d.ts
@@ -1,0 +1,18 @@
+export interface CookieRecord {
+    sid: string;
+    key: string;
+    domain: string;
+    path: string;
+    lastAccessed: Date;
+    syncKey?: string;
+    cookieStr?: string;
+    value?: string;
+    isServerSync?: boolean;
+    isClientSync?: boolean;
+    isWindowSync?: boolean;
+}
+
+export interface ParsedClientSyncCookie {
+    outdated: Array<CookieRecord>;
+    actual: Array<CookieRecord>;
+}
